fix(tdr): break import cycle — extract ticket-readiness module
Move updateSuccessorReadiness + allPredecessorsDone to ticket-readiness.ts
with pool as parameter, removing the circular dependency:
website-db → transition → ticket-graph → website-db.

Also add scs:index:incremental Taskfile task referencing the script
added in the SCS PR (#1566) to satisfy S4 gate.
